refactor output
